pyproject.toml
[project]
name = "dagtor"
dynamic = ["version"]
description = 'Python based orchestrator for DAGs'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "n.michail", email = "nmichail90@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = [
    "returns",
    "podman-compose",
    "psycopg[binary,pool]",
    "databricks-sdk",
    "urllib3<2.0"
] #Add dependencies here

[project.optional-dependencies]
dev = [
    "returns",
    "pytest",
    "pytest-cov",
    "pytest-mock"
]

[project.urls]
Documentation = "https://github.com/n.michail/dagtor#readme"
Issues = "https://github.com/n.michail/dagtor/issues"
Source = "https://github.com/n.michail/dagtor"

[tool.hatch.envs.default]
python = "3.9"
type = "virtual" #This creates a local .venv virtual folder that I can use in Pycharm
path = ".venv"

[tool.hatch.envs.dev]
python = "3.9" 
description = "Custom env that is used for development"
features = ["dev"]
#post-install-commands = ["pre-commit install"]

[build-system]
requires = ["hatchling","hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dagtor"]  # Point directly to the package

[tool.hatch.version]
path = "dagtor/_version.py"
